add integration tests for users

+++ tests/integration/users/test_users_service.py
@@ -1,5 +1,10 @@
+from contextlib import nullcontext
+from typing import ContextManager, Any
+from uuid import uuid4
+
 import pytest
 
+from exceptions import UserNotFoundError
 from schemas import UserSchema, UserSchemaUpdate
 from services.users_service import UsersService
 from .helpers import generate_username
@@ -7,35 +12,87 @@ from ..helpers import get_fake_token_payload
 
 
 @pytest.mark.asyncio
+@pytest.mark.integration
+@pytest.mark.parametrize(
+    ["user_not_found", "expectation"],
+    [
+        (False, nullcontext()),
+        (True, pytest.raises(UserNotFoundError))
+    ]
+)
 async def test_get_user(
         users_service: UsersService,
-        user_data: UserSchema
+        user_data: UserSchema,
+        user_not_found: bool,
+        expectation: ContextManager[Any]
 ) -> None:
-    payload = get_fake_token_payload(user_data.user_id)
+    if user_not_found:
+        payload = get_fake_token_payload(uuid4())
+    else:
+        payload = get_fake_token_payload(user_data.user_id)
 
-    user = await users_service.get_user(payload)
-    assert user.model_dump() == user_data.model_dump()
+    with expectation:
+        user = await users_service.get_user(payload)
+
+    if not user_not_found:
+        assert user.model_dump() == user_data.model_dump()
 
 
 @pytest.mark.asyncio
+@pytest.mark.integration
+@pytest.mark.parametrize(
+    ["user_not_found", "expectation"],
+    [
+        (False, nullcontext()),
+        (True, pytest.raises(UserNotFoundError))
+    ]
+)
 async def test_update_user(
         users_service: UsersService,
-        user_data: UserSchema
+        user_data: UserSchema,
+        user_not_found: bool,
+        expectation: ContextManager[Any]
 ) -> None:
-    paylaod = get_fake_token_payload(user_data.user_id)
+    if user_not_found:
+        payload = get_fake_token_payload(uuid4())
+    else:
+        payload = get_fake_token_payload(user_data.user_id)
+
     user_schema_update = UserSchemaUpdate(username=generate_username())
     user_data.username = user_schema_update.username
 
-    user = await users_service.update_user(paylaod, user_schema_update)
-    assert user.model_dump() == user_data.model_dump()
+    with expectation:
+        user = await users_service.update_user(payload, user_schema_update)
+
+    if not user_not_found:
+        assert user.model_dump() == user_data.model_dump()
 
 
 @pytest.mark.asyncio
+@pytest.mark.integration
+@pytest.mark.parametrize(
+    ["user_not_found", "expectation"],
+    [
+        (False, nullcontext()),
+        (True, pytest.raises(UserNotFoundError))
+    ]
+)
 async def test_delete_user(
         users_service: UsersService,
-        user_data: UserSchema
+        user_data: UserSchema,
+        user_not_found: bool,
+        expectation: ContextManager[Any]
 ) -> None:
-    payload = get_fake_token_payload(user_data.user_id)
+    if user_not_found:
+        payload = get_fake_token_payload(uuid4())
+    else:
+        payload = get_fake_token_payload(user_data.user_id)
 
-    user = await users_service.delete_user(payload)
-    assert user.model_dump() == user_data.model_dump()
+    with expectation:
+        user = await users_service.delete_user(payload)
+
+    if not user_not_found:
+        assert user.model_dump() == user_data.model_dump()
+
+        with pytest.raises(UserNotFoundError):
+            await users_service.delete_user(payload)
